perf: twice faster method check
