+++ lib/core/request.js
@@ -80,7 +80,7 @@ class Request {
 
     if (typeof method !== 'string') {
       throw new InvalidArgumentError('method must be a string')
-    } else if (tokenRegExp.exec(method) === null) {
+    } else if (!util.isValidHTTPToken(method)) {
       throw new InvalidArgumentError('invalid request method')
     }
